pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudstream-telegram-bot"
version = "0.2.0"
description = "Telegram bot that wraps the Asia2TV and AnimeWitcher Cloudstream plugins."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dependencies = [
    "python-telegram-bot>=21.4,<22",
    "httpx>=0.27,<0.29",
    "beautifulsoup4>=4.12,<5",
    "nest_asyncio>=1.6",
]

[project.scripts]
cloudstream-telegram-bot = "cloudstream_bot.bot:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cloudstream_bot = ["i18n/*.json"]
